reads everything until 3box thread. but building slow af
