+++ package.json
@@ -53,6 +53,7 @@
     "react-dom": "^16.13.1",
     "react-draggable": "^4.3.1",
     "react-frame-component": "^4.1.2",
+    "react-stars": "^2.2.5",
     "resolve": "1.15.0",
     "resolve-url-loader": "3.1.1",
     "sass-loader": "8.0.2",
