+++ src/newcomment.js
@@ -1,34 +1,115 @@
-import React, {Component} from 'react';
+import React, { Component } from "react";
+//import ReactStars from 'react-stars';
+import Draggable from 'react-draggable';
 
-export default class AddApp extends Component {
-  render() {
-       return (
-      <section className="py-12 px-4 text-center">
-        {/*<h2 className="text-4xl mb-2 leading-tight font-heading">...so we match these two sites, exactly where they now dont meet. this is how it works...</h2>
-        <p className="text-gray-500">click the div to auth with ethereum and open your first note</p>*/}
-        <div className="flex flex-wrap items-center -mx-8 mt-12 mb-2 font-mono">
-          <div className="lg:w-1/4 px-8 mb-8">
-            <p className=" mb-2">
-            personalized content
-            </p>
-          </div>
-          <div className="lg:w-1/4 px-8 mb-8">
-            <p className=" mb-2">
-            data privacy
-            </p>
-          </div>
-          <div className="lg:w-1/4 px-8 mb-8">
-            <p className=" mb-2">
-            rewards on the blockchain
-            </p>
-          </div>
-          <div className="lg:w-1/4 px-8 mb-8">
-            <p className=" mb-2">
-            interoperable
-            </p>
-          </div>
-        </div>
-      </section>
-    );
+const blankState = {
+  comment: "",
+  rating: "",
+  deltaPosition: {
+        x: 0, y: 0
+      }
+};
+export default class AppForm extends Component {
+  state = blankState;
+
+  handleChange = event => {
+    this.setState(Object.assign({ [event.target.name]: event.target.value }));
+  };
+
+  ratingChanged = (newRating) => {
+    console.log(`${newRating} stars`);
+    this.setState(Object.assign({ rating: newRating }));
+  };
+
+  async validateFormFields() {
+    console.log("to do - validiate form");
   }
+
+  handleSubmit = event => {
+    event.preventDefault();
+    this.validateFormFields();
+
+    this.props.savePost({
+      comment: this.state.comment,
+      rating: this.state.rating,
+      deltaPosition: this.state.deltaPosition,
+    });
+
+    this.setState(Object.assign({}, blankState, { submitted: true }));
+    console.log(this.state.submitted);
+  };
+
+
+  handleDrag = (e, ui, deltaPosition) => {
+      const {x, y} = this.state.deltaPosition;
+      this.setState(Object.assign({ deltaPosition: { x: x + ui.deltaX, y: y + ui.deltaY } }));
+    };
+
+  onControlledDragStop = (e, position) => {
+      const {x, y} = position;
+      this.setState(Object.assign({ deltaPosition: {x, y}}));
+    };
+
+  render() {
+    return (
+      <Draggable
+        //axis="x"
+        //handle=".handle"
+        defaultPosition={{x: 0, y: 0}}
+        position={this.deltaPosition}
+        //grid={[25, 25]}
+        scale={1}
+        onStart={this.handleStart}
+        onDrag={this.handleDrag}
+        onStop={this.onControlledDragStop}
+      >
+          <div className="w-full max-w-sm mx-auto">
+            {/*<div className="handle">Drag here</div>*/}
+            {/*<div>x: {this.state.deltaPosition.x.toFixed(0)}, y: {this.state.deltaPosition.y.toFixed(0)}</div>*/}
+            {!this.state.submitted && (
+              <form onSubmit={this.handleSubmit}>
+                <div className="py-4 flex flex-row items-center">
+                  <div className="mx-4">
+                   How do your rate this site?
+                  </div>
+                  {/*<ReactStars
+                    count={5}
+                    onChange={this.ratingChanged}
+                    size={20}
+                    color2={'#ffd700'}
+                    edit={true}
+                    half={false}
+                    value={this.state.rating}
+                  />*/}
+                </div>
+                <div className="mb-4">
+                  <textarea
+                    className="w-full shadow-inner p-4 border-0"
+                    placeholder="Add your comment."
+                    rows="2"
+                    value={this.state.comment}
+                    onChange={this.handleChange}
+                    type="text"
+                    name="comment"
+                    aria-describedby="commentText"
+                  />
+                </div>
+
+                <div className="mb-6">
+                  <input type="submit" value="Submit" className="inline-block w-full py-4 px-8 leading-none text-white bg-indigo-500 hover:bg-indigo-600 rounded shadow" />
+                </div>
+
+              </form>
+            )}
+            {/*{this.state.submitted && <div className="jumbotron">
+                <h1>Thank you for submiting</h1>
+                <button className="inline-block w-full py-4 px-8 leading-none text-white bg-indigo-500 hover:bg-indigo-600 rounded shadow"  onClick={()=>(this.setState({submitted : false}))}>
+                  Add another comment
+                </button>
+              </div>
+            }*/}
+          </div>
+        </Draggable>
+        );
+      }
 }
